added bricks, model, teleport scirpts snap points
added bricks into the sence. added a model on a display plate on model desk and made it rotate. added snappoints to desk and resource bricks to try snapping in a designated poisition of the desk. wrote respective script for that. implemented return script for making bricks be teleported to resources desk when they fall down. this stage will not include the asset due to licence reasons.
Note: Johanna and John did this push